+++ src/ProjectFileTools.NuGetSearch/IO/WebRequestFactory.cs
@@ -6,16 +6,14 @@ namespace ProjectFileTools.NuGetSearch.IO
 {
     public class WebRequestFactory : IWebRequestFactory
     {
+        private static readonly HttpClient _httpClient = new HttpClient();
         public async Task<string> GetStringAsync(string endpoint, CancellationToken cancellationToken)
         {
             try
             {
-                using (HttpClient client = new HttpClient())
-                {
-                    HttpResponseMessage responseMessage = await client.GetAsync(endpoint, cancellationToken).ConfigureAwait(false);
-                    responseMessage.EnsureSuccessStatusCode();
-                    return await responseMessage.Content.ReadAsStringAsync().ConfigureAwait(false);
-                }
+                HttpResponseMessage responseMessage = await _httpClient.GetAsync(endpoint, cancellationToken).ConfigureAwait(false);
+                responseMessage.EnsureSuccessStatusCode();
+                return await responseMessage.Content.ReadAsStringAsync().ConfigureAwait(false);
             }
             catch
             {