+++ test/ProjectFileTools.NuGetSearch.Tests/UnitTest1.cs
@@ -1,18 +1,18 @@
 using System;
 using System.Threading;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+// using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NuGet.Frameworks;
 using ProjectFileTools.NuGetSearch.Contracts;
 using ProjectFileTools.NuGetSearch.Feeds;
 using ProjectFileTools.NuGetSearch.Feeds.Disk;
 using ProjectFileTools.NuGetSearch.IO;
+using Xunit;
 
 namespace PackageFeedManagerTests
 {
-    [TestClass]
     public class UnitTest1
     {
-        [TestMethod]
+        [Fact(Skip="Just for testing locally")]
         public void TestMethod1()
         {
             //IFileSystem fileSystem = new MockFileSystem();